LayoutTests/editing/smart-lists/inserting-space-after-hyphen-minus-generates-dashed-list-expected.txt

Typing "- " (hyphen-minus, U+002D) at the start of two consecutive lines should auto-convert them into an unordered dashed list, using the hyphen-minus character in the custom list-style-type marker.
| <ul>
|   class="Apple-dash-list"
|   style="list-style-type: "-  ";"
|   <li>
|     "Hello"
|     <br>
|   <li>
|     "World<#selection-caret>"
